+++ src/just_makeit/templates/cmake/CMakeLists_object_core.cmake
@@ -3,7 +3,7 @@ add_library(<<component>>_core OBJECT <<component>>_core.c)
 target_include_directories(
   <<component>>_core PUBLIC ${CMAKE_SOURCE_DIR}/native/inc
                             ${CMAKE_SOURCE_DIR}/native/inc/<<component>>)
-<<extra_link_on_object_core>>
+<<extra_include_dirs_on_object_core>><<extra_link_on_object_core>>
 add_executable(test_<<component>>_core
                ${CMAKE_SOURCE_DIR}/native/tests/test_<<component>>_core.c)
 target_link_libraries(test_<<component>>_core